+++ src/app/components/product/index.js
