+++ src/app/components/product/product.controller.js
