+++ src/app/components/product/product.module.js
