+++ src/app/components/tag/tag.controller.js
