+++ src/app/components/tag/tag.module.js
